# openspec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KientreAAA Web App API",
    "version": "0.1.0",
    "description": "OpenAPI spec for KientreAAA local-first Next.js web app. Covers module execution, agent runs, NotebookLM, local files, Drive helpers, settings, masked Google OAuth, and update endpoints. SSE endpoints are documented with text/event-stream responses. Secrets are never returned raw; examples use placeholders only."
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Default local production port"
    }
  ],
...
  "paths": {
...
    "/api/agent": {
      "post": {
        "tags": ["agent"],
        "summary": "Run agent-mode module and stream steps over SSE",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jobId": { "type": "string" },
                  "task": { "type": "string" },
                  "moduleKey": { "type": "string", "default": "topic" },
                  "history": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": { "type": "string" },
                        "content": { "type": "string" }
                      }
                    }
                  },
                  "settings": { "type": "object", "additionalProperties": true },
                  "sessionId": { "type": "string" }
                },
                "required": ["task"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-Sent Events stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "event: agent_step\ndata: {\"type\":\"assistant\",\"text\":\"Intent: chạy flow — yêu cầu tạo/xử lý tài liệu\"}\n\n"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["agent"],
        "summary": "Cancel a running agent job",
        "parameters": [
          { "name": "jobId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Cancel result" } }
      }
    },
...
  },
...
}